Create Jellyfin_Episode_Title_Card_Rename
Simply put, this script looks at video files in a specified folder with the season number / episode number (S01E01) naming pattern, and matches them to the image file with the similar pattern. It then pulls the name from the video file and renames the image files to match but also adds "-thumb" so the images will  populate as thumbnails on your Jellyfin server.

The regex pattern will match multiple variations of the naming convention together. For instance, S1E4 will match S01E04, or S01E7 will match S01E07.

If you find custom title cards online but they are named rather poorly, this will rename the images to match the corresponding video file and allow for large scale re-naming, almost instantly.

Lastly, this also only works if there is some semblance of the S00E00 pattern in the image and video file's names. If a video file or image does not have the regex pattern, it will be skipped. I also added a line to skip anything named "folder" as Jellyfin likes to use folder.extension for the show posters or season posters.